Added ability to specify security "kinds"
Added annotations to security critical code, and improved tests for and
consolidated security critical code where reasonable.
