+++ src/main/java/com/radixdlt/crypto/Hash.java
@@ -20,6 +20,7 @@ package com.radixdlt.crypto;
 import com.google.common.base.Suppliers;
 import com.google.common.primitives.UnsignedBytes;
 import com.radixdlt.SecurityCritical;
+import com.radixdlt.SecurityCritical.SecurityKind;
 import com.radixdlt.identifiers.EUID;
 import com.radixdlt.utils.Bytes;
 import java.security.SecureRandom;
@@ -28,7 +29,7 @@ import java.util.Comparator;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-@SecurityCritical
+@SecurityCritical(SecurityKind.HASHING)
 public final class Hash implements Comparable<Hash> {
 	private static final Comparator<byte[]> COMPARATOR = UnsignedBytes.lexicographicalComparator();
 	private static final SecureRandom secureRandom = new SecureRandom();
