+++ src/main/java/com/radixdlt/crypto/encryption/Crypt.java
@@ -27,8 +27,12 @@ import org.bouncycastle.crypto.paddings.PaddedBufferedBlockCipher;
 import org.bouncycastle.crypto.params.KeyParameter;
 import org.bouncycastle.crypto.params.ParametersWithIV;
 
+import com.radixdlt.SecurityCritical;
+import com.radixdlt.SecurityCritical.SecurityKind;
+
 import java.util.Arrays;
 
+@SecurityCritical({ SecurityKind.PK_DECRYPT, SecurityKind.PK_ENCRYPT })
 public final class Crypt {
 
 	private Crypt() {
